+++ fxmanifest.lua
@@ -5,7 +5,7 @@ lua54 'yes'
 name 'XS-Phone'
 author 'XyraL'
 description 'iPhone-style smartphone for QBox/QBCore.'
-version '1.0.2'
+version '1.1.0'
 
 dependencies {
     'ox_lib',